refactor(rename): 카멜케이스 -> 대시 형식으로 변경
