+++ frontend/src/App.tsx
@@ -1,10 +1,8 @@
 import {Route, Routes} from "react-router-dom";
 import {Layout, ProtectLayout} from "./components";
-import {Login} from "./pages/login.tsx";
-import {AuthProvider} from "./context/AuthContext.tsx";
+import {Dashboard, Login, Signup} from "./pages";
+import {AuthProvider} from "./store/AuthContext.tsx";
 import {ROOT} from "./utils/constants.ts";
-import {Signup} from "./pages/signup.tsx";
-import {Dashboard} from "./pages/Dashboard.tsx";
 
 function App() {
     return (
