+++ frontend/src/components/common/blur-overlay.tsx
