+++ frontend/src/components/common/error-display.tsx
