+++ frontend/src/components/common/loading-indicator.tsx
@@ -1,0 +1,1 @@
+import styled, {keyframes} from 'styled-components';export const LoadingIndicator = ({message = "잠시만 기다려주세요"}) => {    return (        <LoadingOverlay>            <SpinnerWrapper>                <SpinnerSVG viewBox="0 0 50 50">                    <circle cx="25" cy="25" r="20"></circle>                </SpinnerSVG>            </SpinnerWrapper>            {message && <LoadingText>{message}</LoadingText>}        </LoadingOverlay>    );};const rotate = keyframes`    0% {        transform: rotate(0deg);    }    100% {        transform: rotate(360deg);    }`;const dash = keyframes`    0% {        stroke-dasharray: 1, 150;        stroke-dashoffset: 0;    }    50% {        stroke-dasharray: 90, 150;        stroke-dashoffset: -35;    }    100% {        stroke-dasharray: 90, 150;        stroke-dashoffset: -124;    }`;const LoadingOverlay = styled.div`    top: 0;    left: 0;    width: 100%;    height: 100%;    background-color: rgba(240, 240, 240, 0.7);    backdrop-filter: blur(5px);    -webkit-backdrop-filter: blur(5px);    display: flex;    flex-direction: column;    justify-content: center;    align-items: center;    z-index: 9999;`;const SpinnerWrapper = styled.div`    width: 60px;    height: 60px;    animation: ${rotate} 2s linear infinite; /* 전체 스피너 회전 */`;const SpinnerSVG = styled.svg`    width: 100%;    height: 100%;    circle {        stroke: #3182f6;        stroke-width: 4px;        stroke-linecap: round;        fill: none;        animation: ${dash} 1.5s ease-in-out infinite;    }`;const LoadingText = styled.p`    color: #555;    font-family: 'Spoqa Han Sans Neo', 'Apple SD Gothic Neo', sans-serif;    font-size: 16px;    margin-top: 20px;    font-weight: 500;`;
