+++ frontend/src/components/layout/query-layout.tsx
