+++ frontend/src/features/dashboard/hooks/useDashboardData.ts
@@ -1,1 +1,1 @@
-import {useQuery} from "@tanstack/react-query";import {getDashboardData} from "../api/dashboardApi.ts";export const useDashboardData = () => {    return useQuery({        queryKey: ['dashboard'],        queryFn: getDashboardData,        retry: 3    })}
+import {useQuery} from "@tanstack/react-query";import {getDashboardData} from "../api/dashboard.api.ts";export const useDashboardData = () => {    return useQuery({        queryKey: ['dashboard'],        queryFn: getDashboardData,        retry: 3    })}
